image_generation/requirements.txt
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.0
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
auto-gptq==0.4.2
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chromadb==0.3.22
click==8.1.7
clickhouse-connect==0.6.8
cmake==3.27.2
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dataclasses-json==0.5.14
datasets==2.14.4
debugpy==1.8.0
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
diffusers==0.25.0
dill==0.3.7
diskcache==5.6.1
docx2txt==0.8
duckdb==0.8.1
ecdsa==0.18.0
emoji==2.8.0
et-xmlfile==1.1.0
evaluate==0.4.1
executing==2.0.0
faiss-cpu==1.7.4
Faker==19.3.0
fastapi==0.101.1
fastapi-login==1.9.2
favicon==0.7.0
ffmpy==0.3.1
filelock==3.12.2
filetype==1.2.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.4.0
fschat==0.2.31
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
gradio==3.50.2
gradio_client==0.6.1
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
htbuilder==0.6.1
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.20.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==6.0.1
InstructorEmbedding==1.0.1
ipykernel==6.25.2
ipython==8.12.3
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langchain==0.0.191
linetimer==0.1.5
lit==16.0.6
llama-cpp-python==0.1.66
llvmlite==0.41.1
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==3.0.0
markdown2==2.4.10
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.6
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclassic==0.5.2
nest-asyncio==1.5.6
networkx==3.1
nh3==0.2.14
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.58.1
numexpr==2.8.5
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
onnx==1.15.0
onnxconverter-common==1.14.0
onnxruntime==1.16.3
onnxruntime-gpu==1.16.3
onnxruntime-tools==1.7.0
open-python==1.1.1
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.8.1.78
opencv-python==4.6.0.66
openpyxl==3.1.2
optimum==1.16.1
orjson==3.9.9
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
parso==0.8.3
passlib==1.7.4
pdfminer.six==20221105
peft==0.7.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
posthog==3.0.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyarrow==12.0.1
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-speech-features==0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.1
ray==2.7.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rouge==1.0.1
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.2
scenedetect==0.5.1
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shortuuid==1.0.11
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
st-annotated-text==4.0.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.25.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
tf2onnx==1.16.0
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.36.2
trio==0.21.0
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.10.0
urllib3==1.26.6
uvicorn==0.25.0
uvloop==0.17.0
validators==0.21.2
watchdog==3.0.0
watchfiles==0.19.0
wavedrom==2.0.3.post3
wcwidth==0.2.8
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.7
widgetsnbextension==4.0.5
xmltodict==0.13.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
